refactor: add .clang-tidy

+++ src/kernel/kernel_s32_scalar.cpp
@@ -11,8 +11,8 @@ jit_fn<int32_t> gen_kernel<int32_t, ISA::Scalar>(std::span<const size_t> permuta
         auto output_ptr = x86::rsi;
 
         for (size_t i : std::views::iota(size_t{0}, target_size)) {
-            as.mov(x86::eax, x86::dword_ptr(input_ptr, permutation[i] * sizeof(int32_t)));
-            as.mov(x86::dword_ptr(output_ptr, i * sizeof(int32_t)), x86::eax);
+            as.mov(x86::eax, x86::dword_ptr(input_ptr, static_cast<int>(permutation[i] * sizeof(int32_t))));
+            as.mov(x86::dword_ptr(output_ptr, static_cast<int>(i * sizeof(int32_t))), x86::eax);
         }
 
         as.ret();
